move shared code to query_testing
